Creates protocols for profile screen
