+++ UpcomingMovies/Scenes/Account/Profile/ProfileSection.swift
@@ -1,0 +1,23 @@
+//
+//  ProfileSection.swift
+//  UpcomingMovies
+//
+//  Created by Alonso on 6/27/20.
+//  Copyright © 2020 Alonso. All rights reserved.
+//
+
+import Foundation
+
+enum ProfileSection {
+    /// Section to show the profile user info
+    case accountInfo
+    
+    /// Shows the user's collections like favorites
+    case collections
+    
+    /// Shows the user's created lists
+    case groups
+    
+    /// Shows the sign out Button
+    case signOut
+}
